requirements.txt
opencv-python
ipykernel
matplotlib
ultralytics
label-studio
faiss-cpu
pandas
requests
pillow
torchvision
torch
scikit-learn
transformers
fashion-clip